+++ resources/modules/customizer/src/js/components/sidebar/modules/nodes-obj/MapperObj.js
@@ -1,0 +1,11 @@
+import MapperNode from "../data/MapperNode";
+import Mapper from './../widgets/Mapper';
+import ChangeIcon from "../../../../../svgs/customizer.svg";
+
+export const MapperObj = {
+  title: 'Mapper',
+  name: 'mapper',
+  selectedNode: MapperNode,
+  node: Mapper,
+  icon: ChangeIcon
+}